precision that makes more sense

+++ shaders/lib/common.glsl
@@ -68,7 +68,7 @@ const bool colortex3Clear = false;
 
     const int colortex0Format = RGBA16F;
     const int colortex5Format = R8;
-    const int shadowcolor1Format = RG32F;
+    const int shadowcolor1Format = RG16;
 */
 const bool colortex0Clear = false; // only so we can keep mipmaps from the previous frame
